Refactor GetNodesThatCanBeDown into common module
